+++ .storybook/main.js
@@ -1,4 +1,4 @@
-// const BundleAnalyzerPlugin = require("webpack-bundle-analyzer").BundleAnalyzerPlugin;
+import remarkGfm from 'remark-gfm';
 
 module.exports = {
   stories: ['*.stories.mdx', '*.stories.@(js|jsx|ts|tsx)'],
@@ -6,31 +6,34 @@ module.exports = {
     '@storybook/addon-essentials',
     '@react-theming/storybook-addon',
     'storybook-dark-mode',
-    '@storybook/addon-mdx-gfm',
     '@storybook/addon-styling',
-  ],
-  webpackFinal: async (config, { configType }) => {
-    // analyze webpack bundle size. To enable run storybook with `cross-env analyze=true`
-    // not needed, but helps debug the problem :)
-    // const shouldAnalyze = process.env.analyze === "true";
-    // if (shouldAnalyze) {
-    //     config.plugins.push(
-    //         new BundleAnalyzerPlugin()
-    //     );
-    // }
-
-    // split into more chunks
-    config.optimization = {
-      splitChunks: {
-        chunks: 'all',
-        minSize: 30 * 1024, // 30KB
-        maxSize: 1024 * 1024, // 1MB
+    {
+      name: '@storybook/addon-essentials',
+      options: {
+        mdxPluginOptions: {
+          mdxCompileOptions: {
+            remarkPlugins: [remarkGfm],
+          },
+        },
       },
-
-      minimize: false,
-      minimizer: [],
-    };
-    return config;
+    },
+  ],
+  typescript: {
+    check: true,
+    checkOptions: {},
+    reactDocgen: 'react-docgen-typescript',
+    reactDocgenTypescriptOptions: {
+      shouldExtractLiteralValuesFromEnum: true,
+      propFilter: (prop) =>
+        prop.parent ? !/node_modules/.test(prop.parent.fileName) : true,
+    },
+  },
+  framework: {
+    name: '@storybook/react-vite',
+    options: {},
+  },
+  docs: {
+    autodocs: true,
   },
   framework: {
     name: '@storybook/react-webpack5',