+++ packages/checkbox-field/src/lib/checkbox-field.spec.tsx
@@ -1,0 +1,10 @@
+import { render } from '@testing-library/react';
+
+import CheckboxField from './checkbox-field';
+
+describe('CheckboxField', () => {
+  it('should render successfully', () => {
+    const { baseElement } = render(<CheckboxField />);
+    expect(baseElement).toBeTruthy();
+  });
+});