+++ packages/checkbox/src/index.ts
@@ -1,1 +1,2 @@
 export * from './lib/checkbox';
+export { Checkbox as default } from './lib/checkbox';