+++ packages/checkbox/src/lib/checkbox.tsx
@@ -23,16 +23,23 @@ export interface CheckboxProps extends Omit<MuiCheckboxProps, PropsToOmit> {
 const IndeterminateIcon = styled('span')<{
   colorType: Color;
   variant: 'filled' | 'outlined';
-}>(({ colorType, theme: { palette }, variant }) => {
+  disabled?: boolean;
+}>(({ colorType, theme: { palette }, variant, disabled }) => {
   const isFilled = variant === 'filled';
   const colorVariant = palette[colorType];
+  const backgroundColor = disabled
+    ? palette.grey[20]
+    : isFilled
+    ? colorVariant.main
+    : 'transparent';
 
+  const borderColor = disabled ? palette.grey[20] : colorVariant.main;
   return {
     width: 14,
     height: 14,
-    backgroundColor: isFilled ? colorVariant.main : 'transparent',
+    backgroundColor: backgroundColor,
     borderRadius: '2px',
-    border: `2px solid ${colorVariant.main}`,
+    border: `2px solid ${borderColor}`,
     '&:before': {
       // center content to create line in the middle
       position: 'absolute',
@@ -43,55 +50,54 @@ const IndeterminateIcon = styled('span')<{
       width: 8.89,
       height: 1.78,
       content: '""',
-      backgroundColor: isFilled
-        ? palette.grayScale.supernova
-        : colorVariant.main,
+      backgroundColor: isFilled ? palette.grey[0] : colorVariant.main,
     },
 
-    'input:hover ~ &': {
+    'input:hover:not([disabled]) ~ &': {
       backgroundColor: isFilled ? colorVariant.dark : colorVariant.light,
       borderColor: colorVariant.dark,
 
       '&:before': {
-        backgroundColor: isFilled
-          ? palette.grayScale.supernova
-          : colorVariant.medium,
+        backgroundColor: isFilled ? palette.grey[0] : colorVariant.medium,
       },
     },
 
     'input:disabled ~ &': {
-      backgroundColor: isFilled
-        ? palette.grayScale.spaceStation
-        : 'transparent',
-      border: `2px solid ${palette.grayScale.spaceStation}`,
+      backgroundColor: isFilled ? palette.grey[20] : 'transparent',
+      border: `2px solid ${palette.grey[20]}`,
 
       '&:before': {
-        backgroundColor: isFilled
-          ? palette.grayScale.supernova
-          : palette.grayScale.spaceStation,
+        backgroundColor: isFilled ? palette.grey[0] : palette.grey[20],
       },
     },
   };
 });
 
-const UncheckedIcon = styled(CheckIcon)<{ colorType: Color; error?: boolean }>(
-  ({ colorType, theme: { palette }, error }) => ({
+const UncheckedIcon = styled(CheckIcon)<{
+  colorType: Color;
+  error?: boolean;
+  disabled?: boolean;
+}>(({ colorType, theme: { palette }, error, disabled }) => {
+  const borderColor = error
+    ? palette.error.main
+    : disabled
+    ? palette.grey[20]
+    : palette.grey[35];
+  return {
     width: 14,
     height: 14,
     borderRadius: '2px',
-    border: `2px solid ${
-      error ? palette.error.main : palette.grayScale.spaceStation
-    }`,
+    border: `2px solid ${borderColor}`,
     // Set color of the check icon
     color: 'transparent',
-    'input:hover ~ &': {
+    'input:hover:not([disabled]) ~ &': {
       backgroundColor: palette[colorType].light,
       borderColor: palette[colorType].dark,
       // Set color of the check icon
       color: palette[colorType].medium || palette[colorType].dark,
     },
-  })
-);
+  };
+});
 
 const CheckedIcon = styled(UncheckedIcon)<{
   colorType: Color;
@@ -103,21 +109,21 @@ const CheckedIcon = styled(UncheckedIcon)<{
   const colors = useMemo(() => {
     if (disabled && !isFilled) {
       return {
-        color: palette.grayScale.spaceStation,
-        borderColor: palette.grayScale.spaceStation,
-        backgroundColor: palette.grayScale.supernova,
+        color: palette.grey[20],
+        borderColor: palette.grey[20],
+        backgroundColor: palette.grey[0],
       };
     }
     if (disabled && isFilled) {
       return {
-        color: palette.grayScale.supernova,
-        borderColor: palette.grayScale.spaceStation,
-        backgroundColor: palette.grayScale.spaceStation,
+        color: palette.grey[0],
+        borderColor: palette.grey[20],
+        backgroundColor: palette.grey[20],
       };
     }
     if (isFilled) {
       return {
-        color: palette.grayScale.supernova,
+        color: palette.grey[0],
         borderColor: palette[colorType].main,
         backgroundColor: palette[colorType].main,
       };
@@ -125,7 +131,7 @@ const CheckedIcon = styled(UncheckedIcon)<{
     return {
       color: palette[colorType].main,
       borderColor: palette[colorType].main,
-      backgroundColor: palette.grayScale.supernova,
+      backgroundColor: palette.grey[0],
     };
   }, [colorType, disabled, isFilled, palette]);
 
@@ -134,7 +140,7 @@ const CheckedIcon = styled(UncheckedIcon)<{
     backgroundColor: colors.backgroundColor,
     borderColor: colors.borderColor,
 
-    'input:hover ~ &': {
+    'input:hover:not([disabled]) ~ &': {
       color: colors.color,
       backgroundColor: colors.backgroundColor,
       borderColor: colors.borderColor,
@@ -143,28 +149,36 @@ const CheckedIcon = styled(UncheckedIcon)<{
 });
 
 // TODO: How could we make it more customizable regarding adding new colors?
-export function Checkbox({
+export const Checkbox = ({
   variant = 'filled',
   color = 'primary',
   disabled,
   ...other
-}: CheckboxProps) {
-  return (
-    <MuiCheckbox
-      disableFocusRipple
-      disableRipple
-      disableTouchRipple
-      disabled={disabled}
-      icon={<UncheckedIcon colorType={color} error={color === 'error'} />}
-      checkedIcon={
-        <CheckedIcon colorType={color} variant={variant} disabled={disabled} />
-      }
-      indeterminateIcon={
-        <IndeterminateIcon colorType={color} variant={variant} />
-      }
-      {...other}
-    />
-  );
-}
+}: CheckboxProps) => (
+  <MuiCheckbox
+    disableFocusRipple
+    disableRipple
+    disableTouchRipple
+    disabled={disabled}
+    icon={
+      <UncheckedIcon
+        colorType={color}
+        error={color === 'error'}
+        disabled={disabled}
+      />
+    }
+    checkedIcon={
+      <CheckedIcon colorType={color} variant={variant} disabled={disabled} />
+    }
+    indeterminateIcon={
+      <IndeterminateIcon
+        colorType={color}
+        variant={variant}
+        disabled={disabled}
+      />
+    }
+    {...other}
+  />
+);
 
 export default Checkbox;