+++ packages/divider/src/lib/divider.tsx
@@ -1,0 +1,7 @@
+import MuiDivider, { DividerProps } from '@mui/material/Divider';
+
+const Divider = (props: DividerProps) => <MuiDivider {...props} />;
+
+export { Divider };
+export default Divider;
+export * from '@mui/material/Divider';