+++ packages/input-field/src/index.ts
@@ -1,0 +1,2 @@
+export * from './lib/input-field';
+export * from './lib/input-formik-field';