+++ packages/styles/src/lib/Theme/util.ts
@@ -1,0 +1,9 @@
+export const addTransparency = (
+  color?: string,
+  alpha: number = 0.5
+): string => {
+  const maxAlpha = 255;
+  const alphaHex = Math.ceil(maxAlpha * alpha).toString(16);
+
+  return `${color}${alphaHex}`;
+};