+++ packages/tokens/.storybook/preview.js
@@ -1,3 +1,0 @@
-require('../../../.storybook/preview');
-
-export * from '../../../.storybook/preview';