+++ packages/wizard/.storybook/preview.js
@@ -1,1 +1,0 @@
-require('../../../.storybook/preview');