+++ packages/tokens/CHANGELOG.md
@@ -2,6 +2,47 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.2.2](https://github.com/Novatics/novatics-ui/compare/tokens-0.2.1...tokens-0.2.2) (2023-06-05)
+
+## [0.2.1](https://github.com/Novatics/novatics-ui/compare/tokens-0.2.0...tokens-0.2.1) (2023-05-04)
+
+## [0.2.0](https://github.com/Novatics/novatics-ui/compare/tokens-0.1.0...tokens-0.2.0) (2023-04-27)
+
+
+### Features
+
+* **root:** fix import stories ([636dbb6](https://github.com/Novatics/novatics-ui/commit/636dbb6413892ac79bd5869afe247a0c28dd7db1))
+* **root:** format files and remove jest ([d9465bc](https://github.com/Novatics/novatics-ui/commit/d9465bc1205be35fa970b607b6cb1d05aca4f756))
+* **root:** migrate tokens ([0584e4a](https://github.com/Novatics/novatics-ui/commit/0584e4a4a86914d165ed31f4b6733a1ee548cb6f))
+
+## 0.1.0 (2023-04-27)
+
+
+### Features
+
+* **chip:** fix merge conflicts ([9102b70](https://github.com/Novatics/novatics-ui/commit/9102b70bb133636fd9aeb3e12dfe0b3fbed3d58b))
+* **divider:** implement divider ([cc841e0](https://github.com/Novatics/novatics-ui/commit/cc841e0783e930fbcf714c10b5c237f84ba6466e))
+* **root:** add changes to deploy using github actions ([32ea83b](https://github.com/Novatics/novatics-ui/commit/32ea83b92cd5f28671dcb6a78d85896ed76d5d1e))
+* **root:** adjust generator to add version config ([3e4d33c](https://github.com/Novatics/novatics-ui/commit/3e4d33c02094754a2cf2389d77aa92ea5c1868a5))
+* **root:** setup tokens ([146788f](https://github.com/Novatics/novatics-ui/commit/146788f05ca6dfd05685111d5170bde69c8796db))
+* **tokens:** update colors and typography based on figma ([751c613](https://github.com/Novatics/novatics-ui/commit/751c613b742bd4332fbba29acc8070060a82772e))
+
+
+### Bug Fixes
+
+* **root:** run prittier and adjust pkg name ([ad7a921](https://github.com/Novatics/novatics-ui/commit/ad7a9216557fe1a57aaadd3ab0378211e05371bf))
+* **tokens:** fix tokens build ([b9e167b](https://github.com/Novatics/novatics-ui/commit/b9e167be4149c52ea32bbccf77c7c4b3619c93f9))
+
+## [0.2.1](https://github.com/Novatics/novatics-ui/compare/tokens-0.2.0...tokens-0.2.1) (2023-04-27)
+
+## [0.2.0](https://github.com/Novatics/novatics-ui/compare/tokens-0.1.0...tokens-0.2.0) (2023-04-19)
+
+
+### Features
+
+* **chip:** fix merge conflicts ([9102b70](https://github.com/Novatics/novatics-ui/commit/9102b70bb133636fd9aeb3e12dfe0b3fbed3d58b))
+* **divider:** implement divider ([cc841e0](https://github.com/Novatics/novatics-ui/commit/cc841e0783e930fbcf714c10b5c237f84ba6466e))
+
 ## 0.1.0 (2023-04-06)